% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skygrowth0.R
\name{add_ggGR_sarscov2skygrowth}
\alias{add_ggGR_sarscov2skygrowth}
\title{Combine growth plots for two skygrowth fits, for example comparing s614 G and D clades}
\usage{
add_ggGR_sarscov2skygrowth(
  x,
  x1,
  date_limits = c(as.Date("2020-03-01"), NA),
  ...
)
}
\arguments{
\item{x}{skygrowth1 fit}

\item{x1}{skygrowth1 fit}

\item{...}{additional arguments passed to ggplot}
}
\description{
Combine growth plots for two skygrowth fits, for example comparing s614 G and D clades
}
