% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combineAndSummarize0.R
\name{combine_logs}
\alias{combine_logs}
\title{Combine BEAST log files after removing burnin}
\usage{
combine_logs(logfns, burnProportion = 0.5, ofn = NULL)
}
\arguments{
\item{ofn}{If not NULL will save the combined log to this file name}

\item{bunrProportion}{Will remove this proportion from the beginning of logs}
}
\value{
Data frame with combined logs
}
\description{
Combine BEAST log files after removing burnin
}
