% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combineAndSummarize0.R
\name{combine_logs_and_traj}
\alias{combine_logs_and_traj}
\title{Combine BEAST log AND trajectory files after removing burnin}
\usage{
combine_logs_and_traj(
  logfns,
  trajfns,
  burnProportion = 0.5,
  ntraj = 200,
  ofn = NULL,
  ofntraj = NULL
)
}
\arguments{
\item{ntraj}{This integer number of trajectories will be sampled from each trajectory file}

\item{ofn}{If not NULL will save the combined log to this file name}

\item{ofntraj}{If not NULL will save the combined traj to this file name}

\item{bunrProportion}{Will remove this proportion from the beginning of logs}
}
\value{
List with with combined logs and traj
}
\description{
The 'Sample' column is modified and shared between the combined log and trajectory files so that they can be referenced against eachother
}
