% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combineAndSummarize0.R
\name{combine_traj}
\alias{combine_traj}
\title{Combine PhyDyn trajectory files after removing burnin}
\usage{
combine_traj(trajfns, burnProportion = 0.5, ntraj = 100, ofn = NULL)
}
\arguments{
\item{ntraj}{This integer number of trajectories will be sampled from each trajectory file}

\item{ofn}{If not NULL will save the combined trajectory log to this file name}

\item{bunrProportion}{Will remove this proportion from the beginning of logs}
}
\value{
Data frame with combined trajectories
}
\description{
Combine PhyDyn trajectory files after removing burnin
}
