% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spike.R
\name{compute_spike614genotype}
\alias{compute_spike614genotype}
\title{Determine the spike 614 genotype for the given fasta alignment
NOTE: requires mafft which will be used to realign input data. Requires seqinr package}
\usage{
compute_spike614genotype(algnfn, algnfn2 = "standard_coords.fasta")
}
\arguments{
\item{algnfn}{path to fasta with nucleotide alignment. This will be re-aligned, so no need for standard coordinates}

\item{algnfn2}{path to which realigned sequences will be saved}
}
\value{
A data frame with a column giving the genotype
}
\description{
Determine the spike 614 genotype for the given fasta alignment
NOTE: requires mafft which will be used to realign input data. Requires seqinr package
}
