% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importations.R
\name{compute_timports}
\alias{compute_timports}
\title{Computes the time of importation events using a one or more treedater time trees and maximum parsimony}
\usage{
compute_timports(tds, numpb = 100, ncpu = 6, excludeBefore = 2020)
}
\arguments{
\item{numpb}{A number of parametric bootstrap replicates for each treedater tree.}

\item{ncpu}{Number of cpus to use}

\item{excludeBefore}{Removes any dates before this point (some outlier sequences merge before then (guangdong))}
}
\value{
A data frame that can be used for plotting etc. *weighted_seed_date* shows the mean and CI of the seed date weighted by number of region tips descended from that node. This gives much less weight to later introductions with few descendents.
}
\description{
Time of importations is taken as the midpoing on branches separating an exogenous parent and a regional daughter; the states of nodes are determined by parsimony
}
\examples{
\dontrun{ 
PATT = 'Netherlands'
o1 = region_time_stratified_sample( PATT , n=50, path_to_align=algn, D = D, path_to_save='temp.fasta', nregion = Inf, time_stratify_region=TRUE)
prep_tip_labels_phydyn( 'temp.fasta'
 , regexprs = c(  PATT, PATT  ) 
 , invert_regexpr = c( FALSE, TRUE )
 , demes = c( 'Il'  , 'exog'  )
 , path_to_save = 'temp.fasta'
)
tds = make_starting_trees ( 'temp.fasta', ntres = 20, ncpu = 6, plotout=NULL )
pldf <- compute_timports( tds )
plot.importations( pldf )
X11(); plot.importationsHist( pldf )
}
}
