% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/starttree1.R
\name{cutie_treedater}
\alias{cutie_treedater}
\title{Running treedater on ML trees on HPC or work stations with many cores}
\usage{
cutie_treedater(tr, ntres = 10, threads = 2, ncpu = 5)
}
\arguments{
\item{threads}{number of treedater fits to run in parallel}

\item{ncpu}{number of cpus to use for each treedater fit}
}
\value{
a bunch of treedater trees corresponding to random resolutions of ML tree
}
\description{
Collapses small edges and resolves polytomies randomly many times
}
