% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampleSelection1.R
\name{exog_sampler2}
\alias{exog_sampler2}
\title{Sample exogenous sequence IDs using a time stratified design and including close genetic distance to a regional sample}
\usage{
exog_sampler2(
  md,
  n,
  smallGDpairs,
  region_sample,
  exclusion_rules = list(),
  dedup = FALSE
)
}
\arguments{
\item{md}{data frame with gisaid metadata}

\item{n}{sample size}

\item{smallGDpairs}{A data frame or path to csv containing <ID1 ID2 Distance> such as produced by the 'tn93' tool}

\item{region_sample}{vector of sequence IDs from region}

\item{exclusion_rules}{Optional list of rules of the form c( <meta data column> , <regular expression> ). Where this pattern matches, a sequence will not be included in the sample}

\item{dedup}{If TRUE will not include identical sequences}
}
\description{
Sample exogenous sequence IDs using a time stratified design and including close genetic distance to a regional sample
}
\examples{
\dontrun{
#This will get a sample from King County in Washington and exclude sequences labelled Washington from the exog sample, since we aren't sure if they are from King County or not
ipatt = '^KingCounty$'
epatt = '.*Washington.*'
regiontips = region_sampler1( md, n = 10  , inclusion_rules = list( c('CityOrCounty', ipatt) ))
exogtips = exog_sampler1( md, 20, D, s, exclusion_rules = list( c('CityOrCounty', epatt) )  )
}
}
