% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampleSelection.R
\name{filter_hostAndDates}
\alias{filter_hostAndDates}
\title{Remove samples from non-human hosts and sequences with incomplete date information and sequences with more than 10 per cent missing}
\usage{
filter_hostAndDates(path_to_align, path_to_save = NULL)
}
\arguments{
\item{path_to_save}{Path to save fasta}
}
\value{
DNAbin alignment
}
\description{
Remove samples from non-human hosts and sequences with incomplete date information and sequences with more than 10 per cent missing
}
