% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampleSelection.R
\name{filter_quality0}
\alias{filter_quality0}
\title{Selects a background reference set of sequences with good quality}
\usage{
filter_quality0(
  path_to_align,
  path_to_save = NULL,
  q_threshold = 0.05,
  minEdge = 1/29000/10,
  ...
)
}
\arguments{
\item{path_to_save}{Where to store (as fasta) the filtered alignment}

\item{q_threshold}{Clock outlier threshold}

\item{minEdge}{minimum branch length (substitutions per site) to stabilize clock inference}
}
\value{
Writes a new fasta aligment. Return value is a treedater tree and dnabin
}
\description{
- bat/pangolin/canine
- sequences with many gaps 
- sequences within incomplete dates
- sequences with poor relationship with a molecular clock
}
