% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampleSelection.R
\name{filter_quality1}
\alias{filter_quality1}
\title{Selects a background reference set of sequences with good quality}
\usage{
filter_quality1(
  path_to_align,
  fn_tree = NULL,
  path_to_save = NULL,
  q_threshold = 0.05,
  minEdge = 1/29000/10,
  deduplicate_identical = TRUE,
  ncpu = 6,
  ...
)
}
\arguments{
\item{fn_tree}{Optional path to a maximum likelihood tree OR a ape::phylo. If not provided, will estimate using fasttree.}

\item{path_to_save}{Where to store (as RDS) the filtered alignment etc}

\item{q_threshold}{Clock outlier threshold}

\item{minEdge}{minimum branch length (substitutions per site) to stabilize clock inference}

\item{deduplicate_identical}{if TRUE, will only include one sequence (most recent) from among sets of identical sequences}
}
\value{
Writes a RDS file containing filtered aligment, time tree, fasttree, and sample times. Return value is a list with the same
}
\description{
- sequences with poor relationship with a molecular clock
- optionally deduplicate identical sequences
}
