% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xmlformatting.R
\name{format_xml0}
\alias{format_xml0}
\title{Generates a runnable XML by inserting sequence data, a starting tree, tip dates, and a model start time to a XML skeleton}
\usage{
format_xml0(xmlfn, fastafn, treefn, start = 2020.085, susc_size = NULL)
}
\arguments{
\item{xmlfn}{The file name of the XML skeleton}

\item{fastafn}{The file name of the sequence data in fasta format}

\item{treefn}{The file name of the newick format starting trees.}

\item{start}{The numeric date when the model dynamics (SEIJR) will initiate}

\item{susc_size}{The initial number of susceptible, replaces SUSC_SIZE in xml, also sets scale for exponential size prior; supported in seijr0.1.3_skeleton.xml}
}
\value{
Character string of runnable xml. Individual XMLs for each starting tree will be written to disk
}
\description{
The names of the sequence data and trees must match and must be in the standard form for inferring tip dates. The XML skeleton must have these entries to be overwritten: 
SEQUENCES
START_TREE
DATE_TRAIT
SEIJR_START
SUSC_SIZE
}
\details{
A XML file will be generated for each starting tree provided
}
