% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xmlformatting.R
\name{format_xml_gmrf_exogsir}
\alias{format_xml_gmrf_exogsir}
\title{Generates a runnable XML by inserting sequence data, a starting tree, tip dates, and number of gmrf change points into a XML skeleton}
\usage{
format_xml_gmrf_exogsir(xmlfn, fastafn, treefn, numchange = 3)
}
\arguments{
\item{xmlfn}{The file name of the XML skeleton}

\item{fastafn}{The file name of the sequence data in fasta format}

\item{treefn}{The file name of the newick format starting trees.}

\item{numchange}{integer number of change points for beta(t)}
}
\value{
Character string of runnable xml. Individual XMLs for each starting tree will be written to disk
}
\description{
The names of the sequence data and trees must match and must be in the standard form for inferring tip dates. The XML skeleton must have these entries to be overwritten: 
SEQUENCES
START_TREE
DATE_TRAIT
BETADEFS
DLOGBETADIM
}
\details{
A XML file will be generated for each starting tree provided
}
