% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xmlformatting.R
\name{gen_gmrf_betablock}
\alias{gen_gmrf_betablock}
\title{Assemble a XML fragment that will define beta(t) aas a linear spline with a given number of change points}
\usage{
gen_gmrf_betablock(tend, tstart = 2020.038, numchange = 3)
}
\description{
Assemble a XML fragment that will define beta(t) aas a linear spline with a given number of change points
}
