% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combineAndSummarize2.R
\name{gmrf_exogsir_plot_Rt}
\alias{gmrf_exogsir_plot_Rt}
\title{Plot reproduction number through time from a SEIJR trajectory sample}
\usage{
gmrf_exogsir_plot_Rt(
  trajdf,
  logdf,
  gamma1 = 56.15385,
  date_limits = c(as.Date("2020-02-01"), NA),
  path_to_save = "Rt.png",
  ...
)
}
\arguments{
\item{logdf}{Either a dataframe or a path to rds containing a data frame with posterior logs}

\item{gamma1}{rate of recovery once infectious}

\item{date_limits}{a 2-vector containing bounds for the plotting window. If the upper bound is missing, will use the maximum time in the trajectories}

\item{path_to_save}{Will save a png here}

\item{changePoints}{the time where beta changes}
}
\value{
a list with derived outputs from the trajectories. The first element is a ggplot object if you want to further customize the figure
}
\description{
Plot reproduction number through time from a SEIJR trajectory sample
}
