% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combineAndSummarize2.R
\name{gmrf_exogsir_plot_size}
\alias{gmrf_exogsir_plot_size}
\title{Plot the cumulative infections through time from a SEIJR trajectory sample}
\usage{
gmrf_exogsir_plot_size(
  trajdf,
  case_data = NULL,
  date_limits = c(as.Date("2020-02-01"), NA),
  path_to_save = "size.png",
  log_y_axis = F,
  k = 0.16,
  ...
)
}
\arguments{
\item{case_data}{An optional dataframe containing reported/confirmed cases to be plotted alongside estimates. *Must* contain columns 'Date' and 'Cumulative'. Ensure Date is not a factor or character (see as.Date )}

\item{date_limits}{a 2-vector containing bounds for the plotting window. If the upper bound is missing, will use the maximum time in the trajectories}

\item{path_to_save}{Will save a png here}

\item{log_y_axis}{}

\item{k}{Overdispersion parameter (default 16 per cent, Lloyd Smith 2005)}
}
\value{
a list with derived outputs from the trajectories. The first element is a ggplot object if you want to further customize the figure
}
\description{
Plot the cumulative infections through time from a SEIJR trajectory sample
}
