% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combineAndSummarize1.R
\name{GMRF_SEIJR_plot_Rt}
\alias{GMRF_SEIJR_plot_Rt}
\title{Plot reproduction number through time from a SEIJR trajectory sample}
\usage{
GMRF_SEIJR_plot_Rt(
  trajdf,
  logdf,
  gamma0 = 73,
  gamma1 = 121.667,
  date_limits = c(as.Date("2020-02-01"), NA),
  path_to_save = "Rt.png",
  changePoints = c(2020.123, 2020.161, 2020.199, 2020.238),
  ...
)
}
\arguments{
\item{logdf}{Either a dataframe or a path to rds containing a data frame with posterior logs}

\item{gamma0}{rate of becoming infectious during incubation period}

\item{gamma1}{rate of recovery once infectious}

\item{date_limits}{a 2-vector containing bounds for the plotting window. If the upper bound is missing, will use the maximum time in the trajectories}

\item{path_to_save}{Will save a png here}

\item{changePoints}{the time where beta changes}
}
\value{
a list with derived outputs from the trajectories. The first element is a ggplot object if you want to further customize the figure
}
\description{
Plot reproduction number through time from a SEIJR trajectory sample
}
