% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combineAndSummarize1.R
\name{GMRF_SEIJR_reproduction_number}
\alias{GMRF_SEIJR_reproduction_number}
\title{Compute R0, growth rate and doubling time for the GMRF or SEIJR model (no difference in this case)}
\usage{
GMRF_SEIJR_reproduction_number(X, gamma0 = 73, gamma1 = 121.667, precision = 3)
}
\arguments{
\item{gamma0}{Rate of leaving incubation period ( per capita per year )}

\item{gamma1}{Rate of recovery (per capita per year)}
}
\value{
Data frame with tabulated results and CI
}
\description{
Also prints to the screen a markdown table with the results. This can be copied into reports. 
The tau & p_h parameters _must_ be in the log files. If that's not the case, you can add fixed values like this: X$seir.tau <- 74; X$seir.p_h <- .2
}
