% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mobility0.R
\name{googmobility}
\alias{googmobility}
\title{Compute time series cross correlations and plots using google mobility and R(t) estimates}
\usage{
googmobility(mobdf, J, X, region = "", regionshort = "", rt = NULL)
}
\arguments{
\item{mobdf}{A data frame with google mobility data for the selected region}

\item{J}{Combined trajectory file from PhyDyn analysis}

\item{X}{Combined log file from PhyDyn analysis}

\item{region}{A title for the plots describing the region of analysis}

\item{regionshort}{A short title that will go in file names}

\item{The}{output of a function to compute R(t), such as SEIJR_plot_Rt or gmrf_exogsir_plot_Rt
\dontrun{
# load the mobility data:
mobdf0 = read.csv( system.file( 'extdata/googmob-1may2020.csv', package = 'sarscov2' ) , stringsAs=FALSE)
# extract new york: 
mobdf <- mobdf0[ mobdf0$sub_region_1=='New York' & mobdf0$sub_region_2=='New York County', ]
# make a nice header for the plot
region = 'New York City, USA'
# load the trajectory and log files:
J = readRDS('traj.rds' ) 
X = readRDS('logs.rds' )
# run it: 
o = googmobility( mobdf, J, X, region = 'New York City, USA', regionshort='newyork' , rt = gmrf_exogsir_plot_Rt(J,  X )  )
}}
}
\value{
A list with the tabulated time series and outputs of pika::rolling_corr
}
\description{
This function requires the pika package https://github.com/mrc-ide/pika/
NOTE this function computes the rolling correlation for a range of lag days, but does not currently use this information
The lag data is returned. 
If R(t) is out of sync with mobility, better correlations can be found by shifting the dates of the R(t) data
}
