% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mobility0.R
\name{googmobility_skygrowth}
\alias{googmobility_skygrowth}
\title{Compute time series cross correlations and plots using google mobility and R(t) estimates from the function skygrowth0}
\usage{
googmobility_skygrowth(mobdf, sg0, region = "", regionshort = "")
}
\arguments{
\item{mobdf}{A data frame with google mobility data for the selected region}

\item{sg0}{Skygrowth output from function sarscov2::skygrowth0}

\item{region}{A title for the plots describing the region of analysis}

\item{regionshort}{A short title that will go in file names}
}
\value{
A list with the tabulated time series and outputs of pika::rolling_corr
}
\description{
This function requires the pika package https://github.com/mrc-ide/pika/
NOTE this function computes the rolling correlation for a range of lag days, but does not currently use this information
The lag data is returned. 
If R(t) is out of sync with mobility, better correlations can be found by shifting the dates of the R(t) data
}
\examples{
\dontrun{
# load the mobility data:
mobdf0 = read.csv( system.file( 'extdata/googmob-1may2020.csv', package = 'sarscov2' ) , stringsAs=FALSE)
# extract new york: 
mobdf <- mobdf0[ mobdf0$sub_region_1=='New York' & mobdf0$sub_region_2=='New York County', ]
# make a nice header for the plot
region = 'New York City, USA'
# run it: 
o = googmobility( mobdf, sg0, region = 'New York City, USA', regionshort='newyork'  )
}
}
