% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/starttree1.R
\name{make_starting_tree}
\alias{make_starting_tree}
\title{Generate a starting tree for beast}
\usage{
make_starting_tree(fn, treeofn = "startTree.nwk", ncpu = 4)
}
\arguments{
\item{fn}{File name of input alignment. Must be in current working directory}

\item{treeofn}{File name for output tree (newick)}

\item{ncpu}{Integer number of cores}
}
\value{
Will write the starting tree and returns the treedater tree
}
\description{
This assumes the sequence data are in the format generated by prep_tip_labels_phydyn
Requirements: 
1) Fasta must have labels matching this form hCoV-19/Netherlands/Haarlem_1363688/2020|EPI_ISL_413572|2020-03-01|2020.16393442623|_Il
2) iqtree must be installed and in the environment
}
