% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/starttree1.R
\name{make_starting_trees}
\alias{make_starting_trees}
\title{Make starting trees}
\usage{
make_starting_trees(fastafn, treeoutfn = "startTrees.nwk", ntres = 1, ncpu = 4)
}
\arguments{
\item{fastafn}{File name of sequence data (needed for ML tree estimation)}

\item{treeoutfn}{File name of trees to be written}

\item{ntres}{integer how many start trees to produce? a distinct xml is made for each}

\item{ncpu}{Number of CPUs to use}
}
\value{
Some treedater trees. New start trees are written to disk
}
\description{
This will generate multiple starting trees by ML & treedater. 
Each tree is produced by a different random resolution of polytomies in the ML tree
Sequence names must have sample times included (see prep_tip_labels)
}
