% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treepl.R
\name{mcc_col_tree_plot}
\alias{mcc_col_tree_plot}
\title{plotting tree with background sequences labelled}
\usage{
mcc_col_tree_plot(
  nexfn,
  mostRecentSampleDate,
  country_dictionary,
  internal = "Region",
  style = 1,
  regionDemes = c("Il", "Ih", "E"),
  ofn = "mcc.png"
)
}
\arguments{
\item{nexfn}{Path to nexus file containing annotated MCC tree (output of treeannotator2)}

\item{mostRecentSampleDate}{A character string containign the date of the most recent sample in the form 2020-03-17}

\item{country_dictionary}{data dictionary to convert country names to world bank continents}

\item{internal}{A character string giving the name of the region of interest}

\item{style}{plot style from 1,2 or 3}

\item{regionDemes}{A character vector of deme names which will be coloured red}

\item{ofn}{Output file name of figure}
}
\value{
A ggtree object which can be customized further
}
\description{
Edit of mcc_tree_plot allowing non-region tips to be coloured by continent.
requires nexus input tree, most recent sample date and dictionary converting second element of 
sequence names to continent.
}
