% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treepl.R
\name{mcc_tree_plot}
\alias{mcc_tree_plot}
\title{Plot the maximum clade credibility tree and showing HPD node heights for nodes with >50 per cent node support. Tips sampled from within the specified demes will be coloured red.}
\usage{
mcc_tree_plot(
  nexfn,
  mostRecentSampleDate,
  regionDemes = c("Il", "Ih", "E"),
  ofn = "mcc.png"
)
}
\arguments{
\item{nexfn}{Path to nexus file containing annotated MCC tree (output of treeannotator2)}

\item{mostRecentSampleDate}{A character string containign the date of the most recent sample in the form 2020-03-17}

\item{regionDemes}{A character vector of deme names which will be coloured red}

\item{ofn}{Output file name of figure}
}
\value{
A ggtree object which can be customized further
}
\description{
You will need to combine your tree logs like this:
logcombiner -trees -burnin <integer corresponding to half of sample, eg 5000000> <treelog1> ... <treelog_n>  <combined.trees>
You can use the *tree_combiner_helper* function. See ?tree_combiner_helper
NOTE make sure that tree logs included here only correspond to those beast runs that passed checks in *combine_logs_and_traj* 
treeannotator2 -limit 0.5 -burnin 0 <combined>.trees <outputFileName>.nex
}
