% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importations.R
\name{plot_importations}
\alias{plot_importations}
\title{Plot a errbar of importation times given a data frame from compute_timports}
\usage{
plot_importations(pldf, showDates = TRUE, ...)
}
\arguments{
\item{pldf}{A data frame made by compute_timports}

\item{showDates}{Will plot dates instead of numeric if TRUE}

\item{...}{passed to errbar}
}
\description{
Plot a errbar of importation times given a data frame from compute_timports
}
