% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importations.R
\name{plot_importationsHist}
\alias{plot_importationsHist}
\title{Plot a histogram of importation times given a data frame from compute_timports}
\usage{
plot_importationsHist(
  pldf,
  showDates = TRUE,
  breaks = 10,
  xlab = "Cluster origin",
  main = "",
  ...
)
}
\arguments{
\item{pldf}{A data frame made by compute_timports}

\item{showDates}{Will plot dates instead of numeric if TRUE}

\item{breaks}{}

\item{xlab}{}
}
\description{
Plot a histogram of importation times given a data frame from compute_timports
}
