% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combineAndSummarize0.R
\name{plot_sample_distribution}
\alias{plot_sample_distribution}
\title{Sampling distribution of sequences}
\usage{
plot_sample_distribution(path_to_nex, path_to_save = "sample_distribution.png")
}
\arguments{
\item{path_to_nex}{Path to mcc nexus tree}

\item{path_to_save}{PNG saved here}
}
\description{
Sampling distribution of sequences
}
