% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combineAndSummarize0.R
\name{plot_sample_distribution_spike614}
\alias{plot_sample_distribution_spike614}
\title{Plot sample times for D and G genotypes  on spike 614}
\usage{
plot_sample_distribution_spike614(
  s614,
  path_to_save = "sample_distribution.png"
)
}
\arguments{
\item{s614}{data frame produced by compute_spike614genotype, also element $s614 produced by s614_phylodynamics}
}
\description{
Plot sample times for D and G genotypes  on spike 614
}
