% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampleSelection1.R
\name{prep_tip_labels_seijr}
\alias{prep_tip_labels_seijr}
\title{Prepare an alignment of regional and exogenous sequences with tip labels for SEIJR in PhyDyn}
\usage{
prep_tip_labels_seijr(algnfn, outfn, regiontips, exogtips, metadata)
}
\arguments{
\item{algnfn}{path to gisaid aligment}

\item{outfn}{path to save analysis alignment}

\item{regiontips}{vector of sequence id}

\item{exogtips}{vector of sequence id}

\item{metadata}{data frame with gisaid metadata}
}
\description{
Prepare an alignment of regional and exogenous sequences with tip labels for SEIJR in PhyDyn
}
