% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treepl.R
\name{quick_region_treeplot}
\alias{quick_region_treeplot}
\title{Plot a rooted tree (may be a time tree) with tips highlighting a matching regex}
\usage{
quick_region_treeplot(td, region, maxdate = NULL)
}
\arguments{
\item{td}{A rooted tree, may be a treedater output}

\item{region}{A regular expression matching the tips that you want to highlight}

\item{maxdate}{If a time tree, you can optionally give the date of the most recent sample}
}
\value{
A ggtree plot
}
\description{
Plot a rooted tree (may be a time tree) with tips highlighting a matching regex
}
