% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importations.R
\name{region_clades}
\alias{region_clades}
\title{Extact all clades in given tree descended from nodes with high confidece of being within region (parsimony)}
\usage{
region_clades(td)
}
\description{
Extact all clades in given tree descended from nodes with high confidece of being within region (parsimony)
}
