% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampleSelection1.R
\name{region_sampler1}
\alias{region_sampler1}
\title{Sample sequence IDs from within a region}
\usage{
region_sampler1(
  md,
  n = 50,
  inclusion_rules = list(),
  dedup = TRUE,
  time_stratify = TRUE
)
}
\arguments{
\item{md}{data frame with gisaid metadata}

\item{n}{sample size within region}

\item{inclusion_rules}{Required list of rules of the form c( <meta data column> , <regular expression> ). Where this pattern matches, a sequence will be included in the sample}

\item{dedup}{If TRUE identical sequences will not be included}

\item{time_stratify}{If TRUE, will collect a time stratified sample from within the region instead of simple random sample}
}
\description{
Sample sequence IDs from within a region
}
\examples{
\dontrun{
#This will get a sample from King County in Washington and exclude sequences labelled Washington from the exog sample, since we aren't sure if they are from King County or not
ipatt = '^KingCounty$'
epatt = '.*Washington.*'
regiontips = region_sampler1( md, n = 10  , inclusion_rules = list( c('CityOrCounty', ipatt) ))
exogtips = exog_sampler1( md, 20, D, s, exclusion_rules = list( c('CityOrCounty', epatt) )  )
}
}
