% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampleSelection.R
\name{region_time_stratified_sample}
\alias{region_time_stratified_sample}
\title{Select a random sample from aligment stratified through time 
AND including close distance matches to a subset of sequences from a particular location
All sequences from given location will be included}
\usage{
region_time_stratified_sample(
  region_regex,
  n,
  path_to_align,
  nregion = NULL,
  path_to_save = NULL
)
}
\arguments{
\item{region_regex}{Sample names matching this regular expression will be retained and closest matches also retained}

\item{n}{sample size from outside region}

\item{nregion}{sample size within region; if null will include everything in region}

\item{path_to_save}{Where to store (as fasta) the filtered alignment}

\item{q_threshold}{Clock outlier threshold}

\item{minEdge}{minimum branch length (substitutions per site) to stabilize clock inference}
}
\value{
A DNAbin alignment. Will also save to path_to_save
}
\description{
Select samples based on quantile of sample time distribution. Requires date to be at and of sequence label
}
