% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampleSelection.R
\name{region_time_stratified_sample1}
\alias{region_time_stratified_sample1}
\title{Select a random sample from aligment stratified through time 
AND including close distance matches to a subset of sequences from a particular region
All sequences from given location will be included}
\usage{
region_time_stratified_sample1(
  region_regex,
  n,
  path_to_align,
  smallGDpairs,
  nregion = NULL,
  path_to_save = NULL,
  time_stratify_region = TRUE
)
}
\arguments{
\item{region_regex}{Sample names matching this regular expression will be retained and closest matches also retained}

\item{n}{sample size from outside region}

\item{smallGDpairs}{A data frame or path to csv containing <ID1 ID2 Distance> such as produced by the 'tn93' tool}

\item{nregion}{sample size within region; if null will include everything in region}

\item{path_to_save}{Where to store (as fasta) the filtered alignment}

\item{time_stratify_region}{If TRUE (default) will perform a time stratified sample within region, otherwise will do a simple random sample}

\item{q_threshold}{Clock outlier threshold}

\item{minEdge}{minimum branch length (substitutions per site) to stabilize clock inference}
}
\value{
A DNAbin alignment. Will also save to path_to_save
}
\description{
Select samples based on quantile of sample time distribution. Requires date to be at and of sequence label. Alternatively can do a simple random sample within a region
}
