% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spike.R
\name{s614_phylodynamics}
\alias{s614_phylodynamics}
\title{Run skygrowth and treestructure on clades defined by s614 polymorphism}
\usage{
s614_phylodynamics(tds, s614, ...)
}
\arguments{
\item{tds}{List or multiPhylo of treedater trees or ape::phylo}

\item{s614}{A data frame produced by compute_spike614genotype}

\item{...}{Additional arguments passed to sarscov2::skygrowth1}
}
\value{
A list with skygrowth for G and D, and a list of treestructure tests for each tree
}
\description{
Run skygrowth and treestructure on clades defined by s614 polymorphism
}
\examples{
\dontrun{
library( sarscov2 ) 
 spikedf = compute_spike614genotype( 'algn3.fasta', 'algn4.fasta' )
 # inspect spikedf to make sure it a assigned D or G 
 # some precomputed treedater trees corersponding to algn3.fasta: 
 tds = readRDS('tds.rds' )  
 result = s614_phylodynamics( tds, spikedf)
 print( result$treestructureTests  )
 # make a plot 
 add_ggR_sarscov2skygrowth( result$sgD, result$sgG )
}
}
