% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combineAndSummarize0.R
\name{SEIJR_plot_daily_inf}
\alias{SEIJR_plot_daily_inf}
\title{Plot the daily new infections through time from a SEIJR trajectory sample}
\usage{
SEIJR_plot_daily_inf(
  trajdf,
  logdf,
  case_data = NULL,
  date_limits = c(as.Date("2020-02-01"), NA),
  path_to_save = "daily.png",
  ...
)
}
\arguments{
\item{logdf}{Either a dataframe or a path to rds containing a data frame with posterior logs}

\item{case_data}{An optional dataframe containing reported/confirmed cases to be plotted alongside estimates. *Must* contain columns 'Date' and 'Cumulative'. Ensure Date is not a factor or character (see as.Date )}

\item{date_limits}{a 2-vector containing bounds for the plotting window. If the upper bound is missing, will use the maximum time in the trajectories}

\item{path_to_save}{Will save a png here}
}
\value{
a list with derived outputs from the trajectories. The first element is a ggplot object if you want to further customize the figure
}
\description{
Plot the daily new infections through time from a SEIJR trajectory sample
}
