% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combineAndSummarize0.R
\name{SEIJR_plot_reporting}
\alias{SEIJR_plot_reporting}
\title{Plot reporting rate through time from SEIJR trajectory sample and reported cases}
\usage{
SEIJR_plot_reporting(
  trajdf,
  case_data,
  date_limits = c(as.Date("2020-02-01"), NA),
  path_to_save = "reporting.png",
  errorbar = T,
  ...
)
}
\arguments{
\item{trajdf}{Either a dataframe or a path to rds containing a data frame with a posterior sample of trajectories (see combine_traj)}

\item{case_data}{dataframe containing reported/confirmed cases to be plotted alongside estimates. *Must* contain columns 'Date' and 'Confirmed'. Ensure Date is not a factor or character (see as.Date )}

\item{date_limits}{a 2-vector containing bounds for the plotting window. If the upper bound is missing, will use the maximum time in the trajectories}

\item{path_to_save}{Will save a png here}

\item{errorbar}{T/F plotting of errorbars on each point}
}
\value{
a list with derived outputs from the trajectories. The first element is a ggplot object if you want to further customize the figure
}
\description{
Plot reporting rate through time from SEIJR trajectory sample and reported cases
}
