% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampleSelection.R
\name{time_stratified_sample}
\alias{time_stratified_sample}
\title{Select a random sample from aligment stratified through time}
\usage{
time_stratified_sample(n, path_to_align, path_to_save = NULL)
}
\arguments{
\item{path_to_save}{Where to store (as fasta) the filtered alignment}

\item{q_threshold}{Clock outlier threshold}

\item{minEdge}{minimum branch length (substitutions per site) to stabilize clock inference}
}
\value{
A DNAbin alignment. Will also save to path_to_save
}
\description{
Select samples based on quantile of sample time distribution. Requires date to be at and of sequence label
}
