% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treepl.R
\name{tree_annotator_windows}
\alias{tree_annotator_windows}
\title{Tree annotator in Windows}
\usage{
tree_annotator_windows(
  inputfile = "combined.trees",
  outputfile = "mcc.nex",
  lowMem = FALSE
)
}
\arguments{
\item{lowMem}{Set to TRUE to run treeanotator in low memory mode.}

\item{burnin}{Integer number of MCMC iters to treat as burnin}

\item{fns}{List of tree log files. If not provided, will recursively search the directory for logs}
}
\description{
Tree annotator in Windows
}
