% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treepl.R
\name{tree_combiner_helper}
\alias{tree_combiner_helper}
\title{Wrapper for logcombiner}
\usage{
tree_combiner_helper(
  burnin,
  fns = NULL,
  ofn = "combined.trees",
  resample = NULL
)
}
\arguments{
\item{burnin}{Unix: Integer number of MCMC iters to treat as burnin; In windows this is the percentage burnin as a number}

\item{fns}{List of tree log files. If not provided, will recursively search the directory for logs}

\item{resample}{set this to downsample logs further when combining}
}
\description{
Wrapper for logcombiner
}
