% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combineAndSummarize0.R
\name{write_phylo_metadata_yaml}
\alias{write_phylo_metadata_yaml}
\title{Automatic generation of metadata.yml
You can add more information if you like}
\usage{
write_phylo_metadata_yaml(
  author = "Lily Geidelberg",
  location = "weifang",
  model_version = "seijr0.0.0",
  start_date = "2020-01-24",
  end_date = "2020-02-24",
  sequence_downsampling = "none",
  first_sample = "2020-01-24",
  last_sample = "2020-02-10",
  internal_seq = 20,
  exog_seq = 33,
  tree_estimation_method = "free_tree",
  n_startingtrees = 20,
  init_clock_rate = 0.0015,
  tree_prior = "PhyDyn_SEIR",
  min_P = 0.01,
  Penalty_Agt_Y = 0,
  T0 = 2019.7,
  exog_init_pop = 0.01,
  S = 500,
  b = 15,
  gamma0 = 73,
  gamma1 = 121.667,
  importRate = 5,
  tau = "estimated",
  p_h = "estimated",
  Ne = 0.1,
  clock_rate_prior_lower = 5e-04,
  clock_rate_prior_upper = 0.005,
  PhydynSEIRTreeScaler = T,
  PhydynSEIRTreeRootScaler = T,
  PhydynSEIRUniformOperator = T,
  CoalescentConstantWilsonBalding = T,
  CoalescentConstantSubtreeSlide = T,
  CoalescentConstantNarrow = T,
  CoalescentConstantWide = T,
  ...
)
}
\description{
Automatic generation of metadata.yml
You can add more information if you like
}
