% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampleSelection.R
\name{prep_tip_labels_phydyn}
\alias{prep_tip_labels_phydyn}
\title{Change the name of sequences to recognize numeric time of sampling and deme}
\usage{
prep_tip_labels_phydyn(path_to_align, path_to_save = NULL, deme = "Il")
}
\arguments{
\item{deme}{The name of the deme to add to the end of each label}

\item{path_to_algn}{A DNAbin alignment *or* a system path (type character) where original alignment can be found}
}
\description{
Change the name of sequences to recognize numeric time of sampling and deme
}
