% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combineAndSummarize0.R
\name{SEIJR_reproduction_number}
\alias{SEIJR_reproduction_number}
\title{Compute R0 for the SEIJR.0.0 model}
\usage{
SEIJR_reproduction_number(X, gamma1 = 96, tau = 74, p_h = 0.2)
}
\arguments{
\item{gamma1}{Rate of recovery (per capita per year)}

\item{tau}{Transm rate increase in J}

\item{p_h}{Proportion in J}
}
\value{
Vector of R0 values
}
\description{
Compute R0 for the SEIJR.0.0 model
}
