% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treepl.R
\name{tree_annotator_windows}
\alias{tree_annotator_windows}
\title{Tree annotator in Windows}
\usage{
tree_annotator_windows(inputfile = "combined.trees", outputfile = "mcc.nex")
}
\arguments{
\item{burnin}{Integer number of MCMC iters to treat as burnin}

\item{fns}{List of tree log files. If not provided, will recursively search the directory for logs}
}
\description{
Tree annotator in Windows
}
