% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treepl.R
\name{tree_combiner_helper}
\alias{tree_combiner_helper}
\title{Wrapper for logcombiner}
\usage{
tree_combiner_helper(burnin, fns = NULL, ofn = "combined.trees")
}
\arguments{
\item{burnin}{Integer number of MCMC iters to treat as burnin}

\item{fns}{List of tree log files. If not provided, will recursively search the directory for logs}
}
\description{
Wrapper for logcombiner
}
